+++ components/FormikFields/PaperInputPicker/Looper/index.js
@@ -67,6 +67,9 @@ const Looper = ({
       } else {
         updatedQuestion.formikKey = `${updatedQuestion.formikKey}__sameForm${individualLoopsAdded}`;
       }
+      if (updatedQuestion.options) {
+        updatedQuestion.options = updateTextKeys(updatedQuestion.options);
+      }
       updatedQuestions = updatedQuestions.concat(updatedQuestion);
     });
 
@@ -75,6 +78,17 @@ const Looper = ({
       setLoopsAdded(loopsAdded + 1);
     }
     setIndividualLoopsAdded(individualLoopsAdded + 1);
+  };
+
+  const updateTextKeys = (options) => {
+    let updatedOptions = [];
+    options.forEach((option) => {
+      const updatedOption = _.cloneDeep(option);
+      const textKeys = updatedOption.textKey.split('__');
+      updatedOption.textKey = sameForm ? `__${textKeys[1]}__sameForm${individualLoopsAdded}__${textKeys[2]}` : `__${textKeys[1]}__loop${individualLoopsAdded}__${textKeys[2]}`;
+      updatedOptions = updatedOptions.concat(updatedOption);
+    });
+    return updatedOptions;
   };
 
   const removeLoop = () => {