+++ context/auth.context.js
@@ -1,0 +1,116 @@
+import React, { createContext, useEffect, useState } from 'react';
+
+import { deleteData, getData, storeData } from '../modules/async-storage';
+import {
+  retrieveCurrentUserAsyncFunction,
+  retrieveSignInFunction,
+  retrieveSignOutFunction,
+  retrieveSignUpFunction
+} from '../services/parse/auth/index';
+
+export const UserContext = createContext();
+
+export const UserContextProvider = ({ children }) => {
+  const [isLoading, setIsLoading] = useState(false);
+  const [user, setUser] = useState(null);
+  const [error, setError] = useState(null);
+
+  useEffect(() => {
+    setIsLoading(true);
+    retrieveCurrentUserAsyncFunction().then((currentParseUser) => {
+      if (currentParseUser) {
+        const usr = currentParseUser;
+        usr.isOnline = true;
+        setUser(usr);
+        storeData(usr, 'currentUser');
+        setIsLoading(false);
+      } else {
+        getData('currentUser').then((currentAsyncUser) => {
+          if (currentAsyncUser) {
+            const usr = currentAsyncUser;
+            usr.isOnline = false;
+            setUser(usr);
+          }
+          setIsLoading(false);
+        });
+      }
+    });
+  }, []);
+
+  const onlineLogin = async (enteredCredentials) => {
+    const { username, password } = enteredCredentials;
+    setIsLoading(true);
+    return retrieveSignInFunction(username, password)
+      .then((currentParseUser) => {
+        const usr = currentParseUser;
+        usr.isOnline = true;
+        setUser(usr);
+        storeData(usr, 'currentUser');
+        storeData(password, 'password');
+        setIsLoading(false);
+        return true;
+      })
+      .catch(async (e) => {
+        setError(e.toString());
+        setIsLoading(false);
+        return false;
+      });
+  };
+
+  const offlineLogin = (enteredCredentials) => {
+    const { username, password } = enteredCredentials;
+    const { username: usr, password: pswd } = user; // cached user
+
+    setIsLoading(true);
+
+    if (username !== usr && password !== pswd) return false;
+
+    setIsLoading(false);
+
+    return true;
+  };
+
+  /**
+   * @param {*} params
+   * @returns User Object
+   */
+
+  const register = async (params) => {
+    const { password } = params;
+    storeData(password, 'password');
+    setIsLoading(true);
+    try {
+      const u = await retrieveSignUpFunction(params);
+      setUser(u);
+      setIsLoading(false);
+    } catch (e) {
+      setIsLoading(false);
+      setError(e.toString());
+    }
+  };
+
+  const onLogout = async () => retrieveSignOutFunction()
+    .then(() => {
+      setUser(null);
+      setError(null);
+      deleteData('currentUser');
+      return true;
+    });
+
+  return (
+    <UserContext.Provider
+      value={{
+        isAuthenticated: !!user,
+        user,
+        isLoading,
+        error,
+        onlineLogin,
+        offlineLogin,
+        register,
+        onLogout,
+      }}
+    >
+      {children}
+    </UserContext.Provider>
+  );
+};