+++ domains/Settings/SettingsHome/AccountSettings/OfflineData/config/config.js
@@ -1,0 +1,16 @@
+const configArray = [
+  {
+    label: 'accountSettings.offlineData',
+    fieldType: 'header',
+    formikKey: 'none_offlinedataqueryer',
+  },
+  {
+    label: 'global.commName',
+    formikKey: 'communityname',
+    value: '',
+    fieldType: 'autofillms',
+    parameter: 'CommunitiesUserEntered',
+    validation: false
+  }
+];
+export default configArray;