+++ services/parse/crud/index.js
@@ -1,14 +1,38 @@
 import { Parse } from 'parse/react-native';
 
-import { customMultiParamQueryService, customQueryService } from './custom-queries';
+import { customMultiParamQueryService, customMultiValueArrayService, customQueryService } from './custom-queries';
 
 function retrieveHelloFunction() {
   Parse.Cloud.run('hello').then((result) => result);
 }
 function residentIDQuery(params) {
+  const { parseParam, limit } = params;
+  function checkIfAlreadyExist(accumulator, currentVal) {
+    return accumulator.some((item) => (item.get('fname') === currentVal.get('fname')
+      && item.get('lname') === currentVal.get('lname')
+      && item.get('sex') === currentVal.get('sex')
+      && item.get('marriageStatus') === currentVal.get('marriageStatus')
+      && item.get('educationLevel') === currentVal.get('educationLevel')
+    ));
+  }
+
   return new Promise((resolve, reject) => {
-    Parse.Cloud.run('basicQuery', params).then((result) => {
-      resolve(result);
+    const Model = Parse.Object.extend('SurveyData');
+    const query = new Parse.Query(Model);
+
+    query.descending('createdAt');
+
+    query.equalTo('surveyingOrganization', parseParam);
+    query.limit(limit);
+
+    query.find().then((records) => {
+      const deDuplicatedRecords = records.reduce((accumulator, current) => {
+        if (checkIfAlreadyExist(accumulator, current)) {
+          return accumulator;
+        }
+        return [...accumulator, current];
+      }, []);
+      resolve(JSON.parse(JSON.stringify(deDuplicatedRecords)));
     }, (error) => {
       reject(error);
     });
@@ -68,6 +92,7 @@ function postOfflineForms(params) {
 export {
   countService,
   customMultiParamQueryService,
+  customMultiValueArrayService,
   customQueryService,
   getObjectsByGeolocation,
   postObjectsToClass,